+++ tests/tests.R
@@ -268,6 +268,13 @@ connection && Sys.getenv("NOT_CRAN") == "true" && {
   if (!is_error_value(m[])) test_equal(m[], "Seattle")
   test_class("errorValue", mirai(q(), .timeout = 1000)[])
   test_zero(daemons(0))
+  Sys.sleep(0.5)
+  test_zero(daemons(n = 1L, url = local_url(), dispatcher = TRUE))
+  task <- mirai(TRUE)
+  url <- nextget("url")
+  test_equal(3L, daemon(url = url, maxtasks = 1L, cleanup = 0L, dispatcher = TRUE))
+  test_zero(daemons(n = 0L))
+  test_type("integer", .Random.seed)
 }
 # TLS tests
 connection && Sys.getenv("NOT_CRAN") == "true" && {
@@ -380,14 +387,30 @@ connection && Sys.getenv("NOT_CRAN") == "true" && {
   Sys.setenv(R_DEFAULT_PACKAGES = "stats,utils")
   test_equal(daemons(4), 4L)
   Sys.unsetenv("R_DEFAULT_PACKAGES")
+  test_type("list", everywhere(a <<- TRUE))
+  test_true(all(mirai_map(1:4, function(x) { Sys.sleep(0.2); a})[.flat]))
   for (i in seq_len(10000L)) {q[[i]] <- mirai(1L); attr(q[[i]], "status") <- status()}
   test_equal(sum(unlist(collect_mirai(q))), 10000L)
   test_true(all(as.logical(lapply(lapply(q, attr, "status"), is.list))))
   for (i in seq_len(10000L)) {q[[i]] <- mirai({Sys.sleep(0.001); rnorm(1)}); attr(q[[i]], "status") <- status()}
   test_equal(length(unique(unlist(collect_mirai(q)))), 10000L)
   test_true(all(as.logical(lapply(lapply(q, attr, "status"), is.list))))
-  test_equal(daemons()[["mirai"]][["completed"]], 20000L)
+  test_equal(daemons()[["mirai"]][["completed"]], 20008L)
   test_zero(daemons(0))
+}
+# reproducible RNG tests
+connection && Sys.getenv("NOT_CRAN") == "true" && {
+  test_equal(2L, daemons(2, seed = 1234L))
+  test_equal(1L, launch_local())
+  test_type("character", launch_remote())
+  Sys.sleep(0.5)
+  m <- mirai_map(1:12, rnorm)[]
+  test_zero(daemons(0))
+  Sys.sleep(0.5)
+  test_equal(4L, daemons(4, dispatcher = FALSE, seed = 1234L))
+  n <- mirai_map(1:12, rnorm)[]
+  test_zero(daemons(0))
+  test_identical(m, n)
 }
 test_zero(daemons(0))
 Sys.sleep(1L)